feat(api): stabilize `zks_gasPerPubdata`
