+++ core/lib/web3_decl/src/namespaces/unstable.rs
@@ -7,7 +7,7 @@ use zksync_types::{
         TransactionDetailedResult, TransactionExecutionInfo,
     },
     tee_types::TeeType,
-    L1BatchNumber, L2ChainId, H256, U256,
+    L1BatchNumber, L2ChainId, H256,
 };
 
 use crate::{
@@ -68,7 +68,4 @@ pub trait UnstableNamespace {
         &self,
         tx_bytes: Bytes,
     ) -> RpcResult<TransactionDetailedResult>;
-
-    #[method(name = "gasPerPubdata")]
-    async fn gas_per_pubdata(&self) -> RpcResult<U256>;
 }
