+++ core/node/api_server/src/web3/backend_jsonrpsee/namespaces/unstable.rs
@@ -1,4 +1,3 @@
-use zksync_multivm::zk_evm_latest::ethereum_types::U256;
 use zksync_types::{
     api::{
         ChainAggProof, DataAvailabilityDetails, GatewayMigrationStatus, L1ToL2TxsStatus, TeeProof,
@@ -81,12 +80,6 @@ impl UnstableNamespaceServer for UnstableNamespace {
         tx_bytes: web3::Bytes,
     ) -> RpcResult<TransactionDetailedResult> {
         self.send_raw_transaction_with_detailed_output_impl(tx_bytes)
-            .await
-            .map_err(|err| self.current_method().map_err(err))
-    }
-
-    async fn gas_per_pubdata(&self) -> RpcResult<U256> {
-        self.gas_per_pubdata_impl()
             .await
             .map_err(|err| self.current_method().map_err(err))
     }
